refactored code, moved activities to ui package
