+++ app/src/main/java/com/example/languageflashcard/ui/MainActivity.kt
@@ -1,8 +1,9 @@
-package com.example.languageflashcard
+package com.example.languageflashcard.ui
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.languageflashcard.R
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
